Refactor: 패키지 구조변경을 한다.
