+++ investement/src/main/java/com/example/investment/common/AppConfig.java
@@ -1,4 +1,4 @@
-package com.example.investment.news.common;
+package com.example.investment.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
