+++ investement/src/main/java/com/example/investment/common/RestTemplateClient.java
@@ -1,4 +1,4 @@
-package com.example.investment.template;
+package com.example.investment.common;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpEntity;
